requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
urllib3==2.6.3